Optimize Measurement Service
